[build] move HEADER files into the ouistiti directory
