+++ include/ouistiti/hash.h
