+++ include/ouistiti/utils.h
