+++ src/hash_libb64.c
@@ -28,8 +28,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "httpserver/hash.h"
-#include "httpserver/log.h"
+#include "hash.h"
+#include "log.h"
 
 # include "b64/cencode.h"
 # include "b64/cdecode.h"
